refactor: distance 자료형 변경

+++ src/main/java/subway/domain/path/ShortestPath.java
@@ -7,9 +7,9 @@ import java.util.List;
 public class ShortestPath {
 
     private final List<Station> stations;
-    private final double distance;
+    private final int distance;
 
-    public ShortestPath(List<Station> stations, double distance) {
+    public ShortestPath(List<Station> stations, int distance) {
         this.stations = stations;
         this.distance = distance;
     }
@@ -18,7 +18,7 @@ public class ShortestPath {
         return stations;
     }
 
-    public double getDistance() {
+    public int getDistance() {
         return distance;
     }
 }
